% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_key_bits.R
\name{create_key_bits}
\alias{create_key_bits}
\title{Create ley bits from CBC hashes of a key}
\usage{
create_key_bits(key, image)
}
\arguments{
\item{key}{A text string to use as a seed for the key}

\item{image}{The image that the key will be used with}
}
\value{
A matrix of binary values [0,1] with dimensions equal to the image data
}
\description{
Create ley bits from CBC hashes of a key
}
\examples{
# Assuming we have already created photo using `read_original_image()`
# key_bits = create_key_bits(key = "This_is_the_key", image = photo)
}
