% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_message_image.R
\name{create_message_image}
\alias{create_message_image}
\title{Create an image containing text to be encrypted}
\usage{
create_message_image(
  file = "message.png",
  message = "This is the\\nmessage",
  image
)
}
\arguments{
\item{file}{A file for the PNG to be written to}

\item{message}{The text to inlude in the image}

\item{image}{The matrix containing the original image bits}
}
\value{
A matrix containing pixel data for the message image
}
\description{
Create an image containing text to be encrypted
}
\examples{
photo <- matrix(0, ncol=500, nrow=500)
message_bits <- create_message_image(file = "message.png", message = "This is the\nmessage", photo)
}
