% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_message.R
\name{hide_message}
\alias{hide_message}
\title{Hide a message in a photo using a key}
\usage{
hide_message(image, key, message, crypto_image)
}
\arguments{
\item{image}{The PNG image file you want to hide the message in}

\item{key}{The text to be used as a key to seed the hash}

\item{message}{The plain-text message to hide in the photo}

\item{crypto_image}{The file name to write the photo with the hidden message to}
}
\description{
Hide a message in a photo using a key
}
\examples{
# hide_message(
#   image="John_Tukey.png",
#   key="UseABetterKeyThanThis",
#   message="Put your message here - up to 112 characters",
#   crypto_image="crypto_image.png"
#   )
}
