% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_decrypted_image.R
\name{plot_decrypted_image}
\alias{plot_decrypted_image}
\title{Plot a decrypted message}
\usage{
plot_decrypted_image(plain_text)
}
\arguments{
\item{plain_text}{The matrix of decrypted bits to display}
}
\description{
Plot a decrypted message
}
\examples{
plain_text = matrix(sample(0:1, size=1000, replace=TRUE), ncol=10)
plot_decrypted_image(plain_text)
}
