% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_encrypted_image.R
\name{read_encrypted_image}
\alias{read_encrypted_image}
\title{Read an ecrypted message and decrypt it}
\usage{
read_encrypted_image(file, plain_file = "decrypted.png", key_bits)
}
\arguments{
\item{file}{The file containing the encrypted message}

\item{plain_file}{The file for the decrypted message (as PNG)}

\item{key_bits}{The bits to decrypt the message}
}
\value{
A matrix with the plain text, decrypted message pixel values
}
\description{
Read an ecrypted message and decrypt it
}
\examples{
# decrypted_image = read_encrypted_image(file, plain_file = "decrypted.png", key_bits)
}
