% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_original_image.R
\name{read_original_image}
\alias{read_original_image}
\title{Read a photo into matrix}
\usage{
read_original_image(file = "John_Tukey.png")
}
\arguments{
\item{file}{The photo to be read}
}
\value{
Image pixel values as a matrix
}
\description{
Read a photo into matrix
}
\examples{
# photo = read_original_image(file = "John_Tukey.png")
}
