% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_message.R
\name{reveal_message}
\alias{reveal_message}
\title{Reveal a message hidden in a photo using the key}
\usage{
reveal_message(key, crypto_image)
}
\arguments{
\item{key}{The text to be used as a key to seed the hash}

\item{crypto_image}{The file name of the PNG file with the hidden message}
}
\description{
Reveal a message hidden in a photo using the key
}
\examples{
# reveal_message(
#   key="UseABetterKeyThanThis",
#   crypto_image="crypto_image.png"
#   )
}
