% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_image_layer.R
\name{show_image_layer}
\alias{show_image_layer}
\title{Examine an image at a specific level of significant bits}
\usage{
show_image_layer(file, layer = 1)
}
\arguments{
\item{file}{The PNG image file to examine}

\item{layer}{The layer (bit significance) of the PNG image to view}
}
\value{
Object of class image
}
\description{
Examine an image at a specific level of significant bits
}
\examples{
# show_image_layer(file="John_Tukey.png", layer = 1)
}
