% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_encrypted_image.R
\name{write_encrypted_image}
\alias{write_encrypted_image}
\title{Encrypt and message in a photo}
\usage{
write_encrypted_image(file, message, image, key_bits)
}
\arguments{
\item{file}{The file to write the encrypted photo to}

\item{message}{The message matrix from create_message_image()}

\item{image}{The photo to hide the message in, from read_original_image()}

\item{key_bits}{The key for encryption}
}
\value{
A matrix with the encrypted message hidden in the image pixel values
}
\description{
Encrypt and message in a photo
}
\examples{
# write_encrypted_image(file, message, image, key_bits)
}
